requirements.txt
## conda:
	google-cloud-storage
	google-cloud-bigquery
	pandas
	scikit-learn
	pydot
	python-dotenv

## pip:
	tensorflow
